Refactor package_managers -> sync_npm
Make the naming more consistent between cli, library, and test.
