+++ gratipay/sync_npm/readmes.py
