+++ gratipay/sync_npm/sync.py
@@ -2,14 +2,10 @@
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import argparse
 import csv
 import sys
 import time
 import uuid
-
-from gratipay import wireup
-from gratipay.package_managers import readmes as _readmes
 
 
 log = lambda *a: print(*a, file=sys.stderr)
@@ -134,27 +130,3 @@ def upsert(env, args, db):
                GROUP BY u.package_manager, u.name, u.description, u.emails
 
         """)
-
-
-def fetch_readmes(env, args, db):
-    _readmes.fetch(db)
-
-
-def process_readmes(env, args, db):
-    _readmes.process(db)
-
-
-# cli plumbing
-
-def parse_args(argv):
-    p = argparse.ArgumentParser()
-    p.add_argument('command', choices=['serialize', 'upsert', 'fetch-readmes', 'process-readmes'])
-    p.add_argument('path', help='the path to the input file', nargs='?', default='/dev/stdin')
-    return p.parse_args(argv)
-
-
-def main(argv=sys.argv):
-    env = wireup.env()
-    args = parse_args(argv[1:])
-    db = wireup.db(env)
-    globals()[args.command.replace('-', '_')](env, args, db)
