+++ setup.py
@@ -9,7 +9,7 @@ setup( name='gratipay'
      , entry_points = { 'console_scripts'
                       : [ 'payday=gratipay.cli:payday'
                         , 'fake_data=gratipay.utils.fake_data:main'
-                        , 'sync-npm=gratipay.package_managers.sync:main'
+                        , 'sync-npm=gratipay.sync_npm.cli:main'
                          ]
                        }
       )
